Extract ORESPProvisionManager mock to its own file and use name aligned with other mock
